requirements.txt
spacy>=3.8
google-generativeai>=0.8.5
google-genai>=1.52.0
beautifulsoup4
youtube-transcript-api
requests
shiny
pytest
python-dotenv
